+++ sinch/domains/voice/api/v2/__init__.py
@@ -1,3 +1,5 @@
+from sinch.domains.voice.api.v2.batches_apis import Batches
 from sinch.domains.voice.api.v2.calls_apis import Calls
+from sinch.domains.voice.api.v2.sessions_apis import Sessions
 
-__all__ = ["Calls"]
+__all__ = ["Batches", "Calls", "Sessions"]