+++ sinch/domains/voice/api/v2/voice.py
@@ -1,4 +1,6 @@
+from sinch.domains.voice.api.v2.batches_apis import Batches
 from sinch.domains.voice.api.v2.calls_apis import Calls
+from sinch.domains.voice.api.v2.sessions_apis import Sessions
 
 
 class VoiceV2:
@@ -7,3 +9,5 @@ class VoiceV2:
     def __init__(self, sinch):
         self._sinch = sinch
         self.calls = Calls(self._sinch)
+        self.sessions = Sessions(self._sinch)
+        self.batches = Batches(self._sinch)