+++ sinch/domains/voice/models/v2/shared/__init__.py
@@ -1,8 +1,10 @@
+from sinch.domains.voice.models.v2.shared.call import Call
 from sinch.domains.voice.models.v2.shared.call_destination import (
     CallDestination,
 )
 from sinch.domains.voice.models.v2.shared.call_header import CallHeader
 from sinch.domains.voice.models.v2.shared.call_origin import CallOrigin
+from sinch.domains.voice.models.v2.shared.money import Money
 from sinch.domains.voice.models.v2.shared.phone import Phone, PhoneDetails
 from sinch.domains.voice.models.v2.shared.sip import Sip, SipDetails
 from sinch.domains.voice.models.v2.shared.sip_from import (
@@ -20,9 +22,11 @@ from sinch.domains.voice.models.v2.shared.voice_relay import (
 )
 
 __all__ = [
+    "Call",
     "CallDestination",
     "CallHeader",
     "CallOrigin",
+    "Money",
     "Phone",
     "PhoneDetails",
     "Sip",