+++ sinch/domains/voice/models/v2/types/__init__.py
@@ -1,12 +1,20 @@
 from sinch.domains.voice.models.v2.types.call_destination_dict import (
     CallDestinationDict,
 )
+from sinch.domains.voice.models.v2.types.call_direction import CallDirection
 from sinch.domains.voice.models.v2.types.call_header_dict import CallHeaderDict
 from sinch.domains.voice.models.v2.types.call_origin_dict import CallOriginDict
+from sinch.domains.voice.models.v2.types.call_reason import CallReason
+from sinch.domains.voice.models.v2.types.call_result import CallResult
+from sinch.domains.voice.models.v2.types.call_type import CallType
+from sinch.domains.voice.models.v2.types.origination_type import (
+    OriginationType,
+)
 from sinch.domains.voice.models.v2.types.phone_dict import (
     PhoneDetailsDict,
     PhoneDict,
 )
+from sinch.domains.voice.models.v2.types.session_state import SessionState
 from sinch.domains.voice.models.v2.types.sip_dict import (
     SipDetailsDict,
     SipDict,
@@ -27,10 +35,16 @@ from sinch.domains.voice.models.v2.types.voice_relay_dict import (
 
 __all__ = [
     "CallDestinationDict",
+    "CallDirection",
     "CallHeaderDict",
     "CallOriginDict",
+    "CallReason",
+    "CallResult",
+    "CallType",
+    "OriginationType",
     "PhoneDetailsDict",
     "PhoneDict",
+    "SessionState",
     "SipDetailsDict",
     "SipDict",
     "SipFromDetailsDict",